Pest_Project/requirements.txt
flask
python-dotenv
requests
pillow
flask-cors